c3: complete jsonb bind/return sweeps (postgres dialect green)
Closes out the 71+ JSONB-related compile errors left by the prior
agent's wip commits. Three patterns applied per call site:

* SELECT sites that read JSONB columns into String-typed struct
  fields now cast `col::text AS "col!"` in SQL (preserves struct
  field types for callers that parse downstream).
* INSERT/UPDATE sites binding `&str`/`String` into JSONB columns
  parse to `serde_json::Value` first (rejects malformed JSON early
  with a clear `InvalidData` error).
* Postgres `COUNT(*)` returns nullable bigint; sites that expected
  `i64` now use the `AS "count!: i64"` non-null cast pattern.

Misc collateral:
* `i64`→`i32` for INT columns (tasks.priority, counts, embedding_dim)
* `i64`→`f64` for `interval '1 day' * $n` math
* `mysql_threshold: f64`→`f32` for `ts_rank()` (returns float4)
* `event_taxonomy::text AS "event_taxonomy?"` (nullable column)
* `pr_poller`'s last `?` placeholder → `$1`
* `djinn-agent-worker` and `djinn-server` updated for the
  Postgres-only `DatabaseConnectConfig`/`DatabaseRuntimeConfig` API

+++ server/crates/djinn-db/src/repositories/verification_cache.rs
@@ -69,7 +69,7 @@ impl VerificationCacheRepository {
         self.db.ensure_initialized().await?;
         sqlx::query!(
             r#"DELETE FROM verification_cache WHERE created_at < to_char((now() at time zone 'utc') - (interval '1 day' * $1), 'YYYY-MM-DD"T"HH24:MI:SS.MS"Z"')"#,
-            days,
+            days as f64,
         )
         .execute(self.db.pool())
         .await?;
