+++ server/crates/djinn-db/src/repositories/verification_result.rs
@@ -114,7 +114,7 @@ impl VerificationResultRepository {
         self.db.ensure_initialized().await?;
         sqlx::query!(
             r#"DELETE FROM verification_results WHERE created_at < to_char((now() at time zone 'utc') - (interval '1 day' * $1), 'YYYY-MM-DD"T"HH24:MI:SS.MS"Z"')"#,
-            days,
+            days as f64,
         )
         .execute(self.db.pool())
         .await?;
